pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "vision_suite"
version = "0.1.0"
description = "A workbench of computer vision tools for data preprocessing, training, and inference."
readme = "README.md"
requires-python = ">=3.8"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
    "numpy",
    "opencv-python",
    "Pillow",
    "pycocotools",
    "joblib",
    "tqdm",
    "matplotlib"
]

[project.scripts]
vision-prep = "vision_suite.cli:main"

[tool.setuptools.packages.find]
where = ["src"]
